cleanup js
