+++ assets/scripts/main.js
@@ -1,10 +1,12 @@
 //Import Libraries
-import $ from 'jquery';
+//import $ from 'jquery';
 import Vue from 'vue';
 import { store } from './store';
+import domReady from '@wordpress/dom-ready';
 
 // Import custom modules
 import App from'./modules/app.js';
+import Nav from'./modules/nav.js';
 
 // Import custom components
 import Posts from './components/Posts.vue';
@@ -12,15 +14,16 @@ import Posts from './components/Posts.vue';
 
 //Initiate Classes Here
 const app = new App();
+const nav = new Nav();
 
 
 // Run Class Based Scripts Here
-function run() {
-
-  app.init();
-  app.hamburger(".js-hamburger");
-  app.navigation(".js-trigger-navigation",".js-navigation");
-  app.navigation_sub_menu(".menu-item-has-children");
+domReady( function() {
+//do something after DOM loads.
+  nav.init();
+  nav.hamburger_animate(".js-hamburger");
+  nav.primary(".js-trigger-navigation",".js-navigation");
+  nav.sub_menu(".menu-item-has-children");
 
   if (document.getElementById('#appPosts')){
     const appPosts = new Vue({ 
@@ -32,14 +35,4 @@ function run() {
     });
   }
 
-}
-
-
-// in case the document is already rendered
-if (document.readyState!='loading') run();
-// modern browsers
-else if (document.addEventListener) document.addEventListener('DOMContentLoaded', run);
-// IE <= 8
-else document.attachEvent('onreadystatechange', function(){
-    if (document.readyState=='complete') run();
-});
+} );
